Revamp '_config.yml'
'_config.yml' should only be used for settings that change how Jekyll
works, not to store content data. To this end, a new directory has been
created for storing reusable data:

- '_data/author.yml' contains site author data, like your name, twitter
  handle, etc. These values can be accessed with the liquid tag
  '{{ site.data.author.<key name> }}'

- '_data/site-info.yml' contains info about the site, such as its url,
  title/description, and version number. These values can be accessed
  with the liquid tag '{{ site.data.site-info.<key name> }}'

- To add new types of data, edit an existing file, or add a new
  '<file>.yml'. The new values can be accessed with the liquid tag
  '{{ site.data.<file name>.<key name> }}'

It's a little more verbose, but it's much more organized and prevents
'_config.yml' from becoming too cluttered with mixed sets of data.

---

This commit also fixes a minor issue resulting from these edits that
caused a duplicate '/' character to be added to the front of the
relative urls in '_layouts/post.html', thus breaking them.