Change typeface
PT Sans/Serif are nice, but Source Sans Pro does a bit better for
readability (and has a much smaller file size).

The original link to the PT typefaces is still in the '<head>' in case
you want to revert. There are also Sass variables to make that transition
easier.