Unpublish some posts
Marked a few of the posts as 'published: false' in their YAML front
matter since they're redundant and clutter the index. Turning them "back
on" is as simple as removing the 'published: false' line from their
front matter.

I left them in so you can use them as a reference if needed. They can
ultimately be deleted, if you'd like.